bill-processor/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
imagehash==4.3.1
google-cloud-vision==3.7.2
openai==1.51.0
python-dotenv==1.0.1
pydantic==2.9.2

# Testing
pytest==8.3.3
httpx==0.27.2
